# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Recherche d'emploi",
        "description": "Documentation de l'API Laravel pour la plateforme de recherche d'emploi",
        "contact": {
            "email": "hariznabil663@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur principal"
        }
    ],
    "paths": {
...
        "/api/reservations": {
            "get": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Liste de toutes les réservations (admin)",
                "operationId": "fbc6050d9f7fdb07075d3aaeb157ffa9",
                "responses": {
                    "200": {
                        "description": "Liste récupérée avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Créer une nouvelle réservation",
                "operationId": "01e3d21dc330a31de025203bc142ffa3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "zone_id",
                                    "number_of_tickets"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "zone_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "number_of_tickets": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "total_price": {
                                        "type": "number",
                                        "example": 200.5
                                    },
                                    "payment_status": {
                                        "type": "string",
                                        "example": "unpaid"
                                    },
                                    "reservation_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-12-01T12:00:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Réservation créée avec succès"
                    },
                    "400": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reservations/{id}": {
            "get": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Afficher une réservation",
                "operationId": "a226b43b2a6a846245e658d01bf00720",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la réservation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réservation récupérée avec succès"
                    },
                    "404": {
                        "description": "Réservation non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Mettre à jour une réservation",
                "operationId": "02fa683a9f80a0563596fb06685bcbc7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "number_of_tickets": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "total_price": {
                                        "type": "number",
                                        "example": 300
                                    },
                                    "payment_status": {
                                        "type": "string",
                                        "example": "paid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Réservation mise à jour avec succès"
                    },
                    "404": {
                        "description": "Réservation non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Supprimer une réservation",
                "operationId": "9b6641e4ded1ff7adfce352d589b85e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réservation supprimée avec succès"
                    },
                    "404": {
                        "description": "Réservation non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reservations/search": {
            "get": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Rechercher des réservations",
                "operationId": "02c5fc4c4ed1723a2e522c46b5a3326b",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "zone_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "payment_status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Résultat de la recherche"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reservations/my": {
            "get": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Afficher mes réservations (user)",
                "operationId": "15a5489bd77e1ceb1eadd132c949b69d",
                "responses": {
                    "200": {
                        "description": "Liste des réservations de l'utilisateur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/stripe/webhook": {
            "post": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Webhook Stripe pour créer des réservations après paiement",
                "operationId": "21a9524c0adbd35515c2e7e2825f77d4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook reçu avec succès"
                    },
                    "400": {
                        "description": "Erreur Stripe Webhook"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}